Move workspace page components under page directory
