+++ src/app/pages/workspace/components/footer/workspace-footer.component.html
