+++ src/app/pages/workspace/components/header/workspace-header.component.scss
