+++ src/app/pages/workspace/components/sidebar/workspace-sidebar.component.scss
