+++ src/app/pages/workspace/components/sidebar/workspace-sidebar.component.ts
@@ -1,10 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { Component, ElementRef, Input, ViewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import type { WorkspacePageComponent } from '../../../pages/workspace/workspace.page';
-import { TranslationPipe } from '../../../i18n/translation.pipe';
-import { JsonTreeComponent } from '../../json-tree/json-tree.component';
-import { PageThumbnailsComponent } from '../../page-thumbnails/page-thumbnails.component';
+import type { WorkspacePageComponent } from '../../workspace.page';
+import { TranslationPipe } from '../../../../i18n/translation.pipe';
+import { JsonTreeComponent } from '../../../../components/json-tree/json-tree.component';
+import { PageThumbnailsComponent } from '../../../../components/page-thumbnails/page-thumbnails.component';
 
 @Component({
   selector: 'app-workspace-sidebar',
