+++ src/app/pages/workspace/components/viewer/workspace-viewer.component.scss
